etf_service/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
frozendict==2.4.0
grpcio==1.62.1
grpcio-tools==1.62.1
html5lib==1.1
idna==3.6
lxml==4.9.4
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.1
peewee==3.17.1
protobuf==4.25.3
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-futures==1.0.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
yahooquery==2.3.7
yfinance==0.2.37
